third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-hover-clear.html

<!DOCTYPE html>
<html>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script src="../../../../resources/gesture-util.js"></script>
<script src="../../../../resources/testdriver.js"></script>
<script src="../../../../resources/testdriver-actions.js"></script>
<script src="../../../..//resources/testdriver-vendor.js"></script>

<style type="text/css">
::-webkit-scrollbar {
    width: 0px;
    height: 0px;
}

div {
    line-height: 100px;
}

#hoverme {
    background-color: lightblue;
}

#clickme {
    background-color: lightgreen;
}

#dontclickme {
    background-color: yellow;
    display: none;
}

#hoverme:hover #dontclickme {
    display: block;
}
</style>

<body style="margin:0">

<div id='hoverme'>Hover over me
  <div id='dontclickme'>Don't click me</div>
</div>
<div id='clickme'>Click me</div>

<p>See https://bugs.webkit.org/show_bug.cgi?id=103283 for details.</p>

<script>
const hovered = document.getElementById('hoverme');
const clickMe = document.getElementById('clickme');
const dontClickMe = document.getElementById('dontclickme');
promise_test(async t => {
  assert_equals(dontClickMe.offsetTop, 0, "The Don't Click Me div should not be visible.");
  const hoveredPos = elementCenter(hovered);
  await touchTap(hoveredPos.x, hoveredPos.y);
  assert_equals(dontClickMe.offsetTop, 100, "The Don't Click Me div should be visible after a touchTap on the Hover Over Me div.");

  assert_equals(document.elementFromPoint(250, 250).id, "clickme");
  const clickMePos = elementCenter(clickMe);
  await touchTap(clickMePos.x, clickMePos.y);
  assert_equals(dontClickMe.offsetTop, 0, "The Don't Click Me div should not be visible after a touchTap on the Click Me div.");
}, 'Hover effects from a gesture tap down can be cleared by a gesture tap outside of the hovered element.');

promise_test(async t => {
  assert_equals(dontClickMe.offsetTop, 0, "The Don't Click Me div should not be visible.");
  const hoveredPos = elementCenter(hovered);
  await touchTap(hoveredPos.x, hoveredPos.y);
  assert_equals(dontClickMe.offsetTop, 100, "The Don't Click Me div should be visible after a touchTap on the Hover Over Me div.");

  await mouseMoveTo(350, 350);
  assert_equals(dontClickMe.offsetTop, 0, "The Don't Click Me div should not be visible after a mouse move to below the Click Me div.");
}, 'Hover effects from a gesture tap down can be cleared by a mousemove outside of the hovered element.');
</script>
</body>
</html>
